refactor: move RepoOverview as separate component
